Generate new results on FreeBSD 13.0 with latest tools

+++ SelectSort.java
@@ -10,7 +10,7 @@ import java.io.*;
 
 class Array
 {
-    private int[]  array;
+    private long[]  array;
     private int     array_size;
     
     // Constructor
@@ -33,7 +33,7 @@ class Array
 	{
 	    line = stdin.readLine();
 	    array_size = Integer.parseInt(line);
-	    array = new int[array_size];
+	    array = new long[array_size];
 	}
 	catch(IOException ioex)
 	{
@@ -69,7 +69,7 @@ class Array
 	int     base,
 		c,
 		min;
-	int     temp;
+	long    temp;
 
 	for(base=0; base < array_size-1; base++)
 	{
